+++ Modern JavaScript From The Beginning/Section 8: API Projects/73 - 77. Github Finder/app.js
@@ -19,6 +19,7 @@ searchInput.addEventListener(
                         );
                     } else {
                         ui.showProfile(data.profile);
+                        ui.showRepos(data.repos);
                     }
                 })
                 .catch(error => {